Move update prompt helper local
