+++ client/packages/cli/src/commands/auth/client/update.ts
@@ -2,7 +2,7 @@ import { Effect, Match } from 'effect';
 import type { authClientUpdateDef, OptsFromCommand } from '../../../index.ts';
 import { BadArgsError } from '../../../errors.ts';
 import { GlobalOpts } from '../../../context/globalOpts.ts';
-import { runUIEffect } from '../../../lib/ui.ts';
+import { optOrPrompt, runUIEffect } from '../../../lib/ui.ts';
 import {
   findClientByIdOrName,
   getAppsAuth,
@@ -29,7 +29,6 @@ import {
   hasAnyFlag,
   hasFlag,
   isTrueFlag,
-  optOrPromptIf,
   readPrivateKeyFile,
   redirectSetupMessages,
   redirectUriPrompt,
@@ -153,6 +152,26 @@ const hasGoogleUpdateFlags = (opts: Record<string, unknown>) =>
   isTrueFlag(getFlag(opts, 'dev-credentials')) ||
   hasGoogleCustomCredentialFlags(opts);
 
+const optOrPromptWhenNeeded = (
+  opts: Record<string, unknown>,
+  flag: string,
+  params: {
+    promptIf: boolean;
+    required?: boolean;
+    prompt: UI.TextInputProps;
+  },
+) =>
+  Effect.gen(function* () {
+    const value = getFlag(opts, flag);
+    if (value === undefined && !params.promptIf) return undefined;
+    return yield* optOrPrompt(value, {
+      simpleName: `--${flag}`,
+      required: params.required ?? params.promptIf,
+      skipIf: false,
+      prompt: params.prompt,
+    });
+  });
+
 const selectGoogleUpdateMode = Effect.fn(function* ({
   isWeb,
   switchingFromShared,
@@ -262,7 +281,7 @@ const updateGoogleRedirect = Effect.fn(function* ({
   opts: Record<string, unknown>;
   client: OAuthClientRow;
 }) {
-  const redirectTo = yield* optOrPromptIf(opts, 'custom-redirect-uri', {
+  const redirectTo = yield* optOrPromptWhenNeeded(opts, 'custom-redirect-uri', {
     promptIf: true,
     required: true,
     prompt: newRedirectPrompt,
@@ -315,18 +334,18 @@ const updateGoogleCustomCredentials = Effect.fn(function* ({
   const shouldPromptRedirectUri =
     isWeb && switchingFromShared && promptCredentials;
 
-  const clientId = yield* optOrPromptIf(opts, 'client-id', {
+  const clientId = yield* optOrPromptWhenNeeded(opts, 'client-id', {
     promptIf: shouldPromptClientId,
     required: mustCollectCredentials,
     prompt: clientIdPrompt({ providerUrl: googleConsoleUrl }),
   });
-  const clientSecret = yield* optOrPromptIf(opts, 'client-secret', {
+  const clientSecret = yield* optOrPromptWhenNeeded(opts, 'client-secret', {
     promptIf: shouldPromptClientSecret,
     required: isWeb && mustCollectCredentials,
     prompt: clientSecretPrompt({ providerUrl: googleConsoleUrl }),
   });
   const customRedirectUri = isWeb
-    ? yield* optOrPromptIf(opts, 'custom-redirect-uri', {
+    ? yield* optOrPromptWhenNeeded(opts, 'custom-redirect-uri', {
         promptIf: shouldPromptRedirectUri,
         required: false,
         prompt: redirectPrompt,
@@ -452,17 +471,21 @@ const handleClientIdSecretUpdate = Effect.fn(function* (params: {
     promptRedirect = action === 'redirect';
   }
 
-  const clientId = yield* optOrPromptIf(params.opts, 'client-id', {
+  const clientId = yield* optOrPromptWhenNeeded(params.opts, 'client-id', {
     promptIf: promptCredentials,
     required: promptCredentials,
     prompt: clientIdPrompt({ providerUrl: params.providerUrl }),
   });
-  const clientSecret = yield* optOrPromptIf(params.opts, 'client-secret', {
-    promptIf: promptCredentials,
-    required: promptCredentials,
-    prompt: clientSecretPrompt({ providerUrl: params.providerUrl }),
-  });
-  const customRedirectUri = yield* optOrPromptIf(
+  const clientSecret = yield* optOrPromptWhenNeeded(
+    params.opts,
+    'client-secret',
+    {
+      promptIf: promptCredentials,
+      required: promptCredentials,
+      prompt: clientSecretPrompt({ providerUrl: params.providerUrl }),
+    },
+  );
+  const customRedirectUri = yield* optOrPromptWhenNeeded(
     params.opts,
     'custom-redirect-uri',
     {
@@ -580,29 +603,37 @@ const readAppleWebUpdate = Effect.fn(function* ({
   client: OAuthClientRow;
   promptAll: boolean;
 }) {
-  const teamId = yield* optOrPromptIf(opts, 'team-id', {
+  const teamId = yield* optOrPromptWhenNeeded(opts, 'team-id', {
     promptIf: promptAll,
     required: promptAll,
     prompt: appleTeamIdPrompt({}),
   });
-  const keyId = yield* optOrPromptIf(opts, 'key-id', {
+  const keyId = yield* optOrPromptWhenNeeded(opts, 'key-id', {
     promptIf: promptAll,
     required: promptAll,
     prompt: appleKeyIdPrompt({}),
   });
-  const privateKeyPath = yield* optOrPromptIf(opts, 'private-key-file', {
-    promptIf: promptAll,
-    required: promptAll,
-    prompt: applePrivateKeyFilePrompt({}),
-  });
+  const privateKeyPath = yield* optOrPromptWhenNeeded(
+    opts,
+    'private-key-file',
+    {
+      promptIf: promptAll,
+      required: promptAll,
+      prompt: applePrivateKeyFilePrompt({}),
+    },
+  );
   const privateKey = privateKeyPath
     ? yield* readPrivateKeyFile(privateKeyPath)
     : undefined;
-  const customRedirectUri = yield* optOrPromptIf(opts, 'custom-redirect-uri', {
-    promptIf: promptAll,
-    required: false,
-    prompt: redirectPrompt,
-  });
+  const customRedirectUri = yield* optOrPromptWhenNeeded(
+    opts,
+    'custom-redirect-uri',
+    {
+      promptIf: promptAll,
+      required: false,
+      prompt: redirectPrompt,
+    },
+  );
 
   const meta: Record<string, string> = {};
   if (teamId) meta.teamId = teamId;
@@ -629,7 +660,7 @@ const handleAppleUpdate = Effect.fn(function* (
     yes,
   });
 
-  const servicesId = yield* optOrPromptIf(opts, 'services-id', {
+  const servicesId = yield* optOrPromptWhenNeeded(opts, 'services-id', {
     promptIf: promptAll,
     required: promptAll,
     prompt: appleServicesIdPrompt({}),
@@ -674,7 +705,7 @@ const handleClerkUpdate = Effect.fn(function* (
   client: OAuthClientRow,
 ) {
   const { yes } = yield* GlobalOpts;
-  const publishableKey = yield* optOrPromptIf(opts, 'publishable-key', {
+  const publishableKey = yield* optOrPromptWhenNeeded(opts, 'publishable-key', {
     promptIf: !yes && !hasFlag(opts, 'publishable-key'),
     required: true,
     prompt: clerkPublishableKeyPrompt({}),
@@ -716,7 +747,7 @@ const handleFirebaseUpdate = Effect.fn(function* (
   client: OAuthClientRow,
 ) {
   const { yes } = yield* GlobalOpts;
-  const projectId = yield* optOrPromptIf(opts, 'project-id', {
+  const projectId = yield* optOrPromptWhenNeeded(opts, 'project-id', {
     promptIf: !yes && !hasFlag(opts, 'project-id'),
     required: true,
     prompt: firebaseProjectIdPrompt({}),
